+++ lib/helpers/app_icons.dart
@@ -1,4 +1,4 @@
-/// Flutter icons MyFlutterApp
+/// Flutter icons RelicIcons
 /// Copyright (C) 2020 by original authors @ fluttericon.com, fontello.com
 /// This font was generated by FlutterIcon.com, which is derived from Fontello.
 ///
@@ -7,18 +7,18 @@
 ///
 /// flutter:
 ///   fonts:
-///    - family:  MyFlutterApp
+///    - family:  RelicIcons
 ///      fonts:
-///       - asset: fonts/MyFlutterApp.ttf
+///       - asset: fonts/RelicIcons.ttf
 ///
 /// 
 ///
 import 'package:flutter/widgets.dart';
 
-class MyFlutterApp {
-  MyFlutterApp._();
+class RelicIcons {
+  RelicIcons._();
 
-  static const _kFontFam = 'MyFlutterApp';
+  static const _kFontFam = 'RelicIcons';
   static const _kFontPkg = null;
 
   static const IconData search = IconData(0xe800, fontFamily: _kFontFam, fontPackage: _kFontPkg);
