+++ lib/main.dart
@@ -1,11 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
-import 'package:retro_shopping/screens/cart.dart';
-import 'package:retro_shopping/screens/home.dart';
-import 'package:retro_shopping/screens/search.dart';
-import 'package:retro_shopping/splash.dart';
-import 'screens/profile.dart';
-import 'bottom_nav_bar.dart';
+import 'package:retro_shopping/helpers/constants.dart';
+import 'package:retro_shopping/widgets/splash.dart';
+import 'dashboard.dart';
 
 void main() {
   runApp(MyApp());
@@ -19,67 +16,13 @@ class MyApp extends StatelessWidget {
       debugShowCheckedModeBanner: false,
       theme: ThemeData(
           // primarySwatch: Colors.blue,
-          scaffoldBackgroundColor: const Color(0xff009d9d),
+          scaffoldBackgroundColor: RelicColors.backgroundColor,
           visualDensity: VisualDensity.adaptivePlatformDensity,
           textTheme: GoogleFonts.poppinsTextTheme(Theme.of(context).textTheme)),
       routes: <String, WidgetBuilder>{
         '/dashboard': (BuildContext context) => Dashboard(),
       },
       home: Splash(),
-    );
-  }
-}
-
-class Dashboard extends StatefulWidget {
-  @override
-  _DashboardState createState() => _DashboardState();
-}
-
-class _DashboardState extends State<Dashboard> {
-  PageController _pageController;
-  int _currentIndex = 0;
-
-  @override
-  void initState() {
-    super.initState();
-    _pageController = PageController();
-  }
-
-  @override
-  void dispose() {
-    super.dispose();
-    _pageController.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      resizeToAvoidBottomInset: false,
-      body: Container(
-        height: MediaQuery.of(context).size.height,
-        width: MediaQuery.of(context).size.width,
-        child: Stack(
-          children: <Widget>[
-            PageView(
-              controller: _pageController,
-              onPageChanged: (int index) {
-                setState(() {
-                  _currentIndex = index;
-                });
-              },
-              children: [
-                Home(),
-                Search(),
-                Cart(),
-                ProfilePage(),
-              ],
-            ),
-            Align(
-                alignment: Alignment.bottomCenter,
-                child: FloatingNavBar(_currentIndex, _pageController)),
-          ],
-        ),
-      ),
     );
   }
 }
